LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/maxlength-expected.txt

Text input element


FAIL Unset maxlength is -1 assert_equals: expected -1 but got 524288
FAIL Negative maxlength is always -1 assert_equals: expected -1 but got 524288
FAIL Non-numeric maxlength is -1 assert_equals: expected -1 but got 524288
PASS Assigning negative integer throws IndexSizeError 
PASS Assigning non-numeric to maxlength sets maxlength to 0 
